@@ -1,9 +1,17 @@
+/*
+ * cpx_data.dart
+ * CPX Research
+ *
+ * Created by Dennis Kossmann on 7.7.2021.
+ * Copyright © 2021. All rights reserved.
+ */
 
+import 'package:cpx_research_sdk_flutter/cpx.dart';
+import 'package:cpx_research_sdk_flutter/widgets/cpx_browser_view.dart';
 import 'package:flutter/material.dart';
 
-import 'cpx_controller.dart';
 import 'model/cpx_response.dart';
-import 'utils/network_service.dart';
+import 'utils/cpx_network_service.dart';
 
 /// The class [CPXData] provides a ValueNotifier for surveys and transactions
 class CPXData {
@@ -11,44 +19,44 @@ class CPXData {
   ValueNotifier<List<Survey>?> surveys = ValueNotifier([]);
   ValueNotifier<List<Transaction>?> transactions = ValueNotifier([]);
   ValueNotifier<CPXText?> text = ValueNotifier(null);
-
-  void setSurveys(List<Survey>? surveys) {
-    this.surveys.value = surveys;
-  }
-
-  void setTransactions(List<Transaction>? transactions) {
-    this.transactions.value = transactions;
-  }
-
-  void setText(CPXText? text) {
-    this.text.value = text;
-  }
 }
 
 /// The function [showCpxLayer] shows the CPX Layer
-void showCpxLayer() {
-  Controller.controller.showCPXLayer();
-}
+void showCpxLayer() => CPXController.controller.showCPXLayer();
 
 /// The function [hideCpxLayer] hides the CPX Layer
-void hideCpxLayer() {
-  Controller.controller.hideCPXLayer();
-}
+void hideCpxLayer() => CPXController.controller.hideCPXLayer();
 
-/// The function [showBrowser] shows the CPX Browser
+/// The function [showCPXBrowserOverlay] shows the CPX Browser
 ///
 /// To open the CPX Browser with a specific survey, provide the optional [surveyID].
-///
-void showBrowser([String? surveyID]) {
-  surveyID != null ? Controller.controller.showBrowser(singleSurvey: true, surveyID: surveyID) : Controller.controller.showBrowser();
+void showCPXBrowserOverlay([String? surveyID]) => surveyID != null
+    ? CPXController.controller
+        .showBrowser(singleSurvey: true, surveyID: surveyID)
+    : CPXController.controller.showBrowser();
+
+void showCPXBrowserDialog({
+  required BuildContext context,
+  required CPXConfig config,
+}) {
+  CPXController.controller.config = config;
+  CPXNetworkService().fetchSurveysAndTransactions();
+
+  showDialog(
+    context: context,
+    builder: (context) => SizedBox.expand(
+      child: CPXBrowserView(
+        CPXBrowserTab.home,
+        onClose: () => Navigator.pop(context),
+      ),
+    ),
+  );
 }
 
 /// The function [fetchCPXSurveysAndTransactions] provides surveys and transactions via the survey and transaction listeners
-void fetchCPXSurveysAndTransactions() {
-  NetworkService().fetchSurveysAndTransactions();
-}
+void fetchCPXSurveysAndTransactions() =>
+    CPXNetworkService().fetchSurveysAndTransactions();
 
 /// The function [markTransactionAsPaid] sends an api request and marks transactions as paid
-void markTransactionAsPaid(String transactionID, String messageID) {
-  NetworkService().setTransactionPaid(transactionID, messageID);
-}
+void markTransactionAsPaid(String transactionID, String messageID) =>
+    CPXNetworkService().setTransactionPaid(transactionID, messageID);
