@@ -1,15 +1,23 @@
+/*
+ * cpx_research.dart
+ * CPX Research
+ *
+ * Created by Dennis Kossmann on 7.7.2021.
+ * Copyright © 2021. All rights reserved.
+ */
+
 import 'dart:async';
 
-import 'package:cpx_research_sdk_flutter/widgets/browser_view.dart';
+import 'package:cpx_research_sdk_flutter/widgets/cpx_browser_view.dart';
 import 'package:cpx_research_sdk_flutter/cpx_controller.dart';
 import 'package:cpx_research_sdk_flutter/utils/cpx_logger.dart';
-import 'package:cpx_research_sdk_flutter/utils/network_service.dart';
+import 'package:cpx_research_sdk_flutter/utils/cpx_network_service.dart';
 import 'package:flutter/material.dart';
 
 import 'model/cpx_config.dart';
-import 'widgets/corner.dart';
-import 'widgets/notification.dart';
-import 'widgets/sidebar.dart';
+import 'widgets/cpx_corner.dart';
+import 'widgets/cpx_notification.dart';
+import 'widgets/cpx_sidebar.dart';
 
 class CPXResearch extends StatefulWidget {
   final CPXConfig config;
@@ -28,12 +36,12 @@ class _CPXResearchState extends State<CPXResearch> with WidgetsBindingObserver {
     _timer = Timer.periodic(
       Duration(seconds: 120),
       (Timer t) {
-        NetworkService().fetchSurveysAndTransactions();
-        Controller.controller.showNotification();
+        CPXNetworkService().fetchSurveysAndTransactions();
+        CPXController.controller.showNotification();
       },
     );
     CPXLogger.log("Timer for automatic survey check started");
-    NetworkService().fetchSurveysAndTransactions();
+    CPXNetworkService().fetchSurveysAndTransactions();
   }
 
   /// The function [_stopTimer] stops the automatic survey check
@@ -46,7 +54,7 @@ class _CPXResearchState extends State<CPXResearch> with WidgetsBindingObserver {
   void initState() {
     super.initState();
     WidgetsBinding.instance.addObserver(this);
-    Controller.controller.config = widget.config;
+    CPXController.controller.config = widget.config;
     _startTimer();
   }
 
@@ -56,12 +64,10 @@ class _CPXResearchState extends State<CPXResearch> with WidgetsBindingObserver {
     switch (state) {
       case AppLifecycleState.resumed:
         _startTimer();
-        break;
       case AppLifecycleState.inactive:
         _stopTimer();
-        break;
       default:
-        break;
+        return;
     }
   }
 
@@ -73,37 +79,40 @@ class _CPXResearchState extends State<CPXResearch> with WidgetsBindingObserver {
   }
 
   @override
-  Widget build(BuildContext context) {
-    return ValueListenableBuilder<bool>(
-      valueListenable: Controller.controller.isCPXResearchLayerDisplayed,
-      builder: (context, isCPXResearchLayerDisplayed, _) {
-        if (isCPXResearchLayerDisplayed) {
-          return Container(
-            height: MediaQuery.of(context).size.height,
-            width: MediaQuery.of(context).size.width,
-            child: ValueListenableBuilder<bool>(
-              valueListenable: Controller.controller.areCPXWidgetsDisplayed,
-              builder: (context, areCPXWidgetsDisplayed, _) {
-                if (areCPXWidgetsDisplayed) {
-                  return SafeArea(
-                    child: Stack(
-                      children: [
-                        if (widget.config.cornerWidget != null) Corner(widget.config.cornerWidget!),
-                        if (widget.config.sidebarWidget != null) Sidebar(widget.config.sidebarWidget!),
-                        if (widget.config.notificationWidget != null) NotificationWidget(widget.config.notificationWidget!)
-                      ],
+  Widget build(BuildContext context) => ValueListenableBuilder<bool>(
+        valueListenable: CPXController.controller.isCPXResearchLayerDisplayed,
+        builder: (context, isCPXResearchLayerDisplayed, _) =>
+            isCPXResearchLayerDisplayed
+                ? Container(
+                    height: MediaQuery.of(context).size.height,
+                    width: MediaQuery.of(context).size.width,
+                    child: ValueListenableBuilder<bool>(
+                      valueListenable:
+                          CPXController.controller.areCPXWidgetsDisplayed,
+                      builder: (context, areCPXWidgetsDisplayed, _) =>
+                          areCPXWidgetsDisplayed
+                              ? SafeArea(
+                                  child: Stack(
+                                    children: [
+                                      if (widget.config.cornerWidget != null)
+                                        CPXCorner(widget.config.cornerWidget!),
+                                      if (widget.config.sidebarWidget != null)
+                                        CPXSidebar(
+                                            widget.config.sidebarWidget!),
+                                      if (widget.config.notificationWidget !=
+                                          null)
+                                        CPXNotification(
+                                            widget.config.notificationWidget!)
+                                    ],
+                                  ),
+                                )
+                              : Scaffold(
+                                  backgroundColor: Colors.black38,
+                                  body: CPXBrowserView(
+                                      CPXController.controller.activeTab),
+                                ),
                     ),
-                  );
-                } else {
-                  return BrowserView(Controller.controller.activeTab);
-                }
-              },
-            ),
-          );
-        } else {
-          return Container();
-        }
-      },
-    );
-  }
+                  )
+                : const SizedBox(),
+      );
 }
