@@ -1,4 +1,12 @@
 
+/*
+ * cpx_config.dart
+ * CPX Research
+ *
+ * Created by Dennis Kossmann on 7.7.2021.
+ * Copyright © 2021. All rights reserved.
+ */
+
 import 'package:flutter/material.dart';
 
 import 'cpx_style.dart';
@@ -9,7 +17,7 @@ import 'cpx_style.dart';
 ///
 /// The [userID] is your external user id
 ///
-/// The [singleSurvey] defines whether the webview is showing all or only a single survey
+/// The [accentColor] is the color of the widgets
 ///
 /// The [cornerWidget] defines the style of the corner widget. If you don't want to display the widget leave it out.
 ///
